+++ CHANGELOG.md
@@ -5,12 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Added cache mechanism in DiscoveryFinderClientImpl for findDiscoveryEndpoints
+
+## [4.3.0] - 2023-12-08
+### Added
+- Added support for `hasAlternatives` property in SingleLevelBomAsBuilt aspect
+- Added EDR token cache to reuse token after contract negotiation
+
 ### Changed
 - Updated edc dependencies to 0.2.1
 - Update deprecated field `providerUrl` to `counterPartyAddress` in EDC catalog request
 - Update ESS EDC notification creation asset endpoint to v3
-- Added EDR token cache to reuse token after contract negotiation
-
 
 ## [4.2.0] - 2023-11-28
 ### Changed
@@ -474,7 +482,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Unresolved
 - **Select Aspects you need**  You are able to select the needed aspects for which you want to collect the correct endpoint information.
 
-[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.2.0...HEAD
+[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.3.0...HEAD
+[4.3.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.2.0...4.3.0
 [4.2.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.1.0...4.2.0
 [4.1.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.0.2...4.1.0
 [4.0.2]: https://github.com/eclipse-tractusx/item-relationship-service/compare/4.0.1...4.0.2